Fix for #132, part 2. Number of spaces before opening bracket not checked for footnotes.
The issue was just that, that the number of spaces were not checked.
The fix was more complicated, though, because to check the number of spaces at the beginning of the line, one must make sure to not consume them when looking for the previous line ending.
The solution was to make the parsing more strict. As a consequence, the backtracking logic could be simplified a bit.